feat: Attraction controller fix: no file saving when adding pic to blob storage; service at startup

+++ TravelRecordsAPI/TravelRecordsAPI/Controllers/StagesController.cs
@@ -66,7 +66,7 @@ namespace TravelRecordsAPI.Controllers
             }
             catch (DbUpdateException)
             {
-                if (StageExists(stage.TripId))
+                if (StageExists((int)stage.TripId))
                 {
                     return Conflict();
                 }
